Extract `Connect` component to `IdentityContext`
