+++ libs/web/components/src/IdentityForm.tsx
@@ -1,14 +1,8 @@
-import { AnyTuple } from "@cennznet/types";
-import type { Option } from "@polkadot/types-codec";
-import type { IdentityInfo } from "@polkadot/types/interfaces";
-import { FC, useContext, useEffect, useState } from "react";
-import { classNames, If } from "react-extras";
+import { FC, useContext } from "react";
 
-import { useCENNZApi, useCENNZWallet } from "@app-gov/web/providers";
 import { IntrinsicElements } from "@app-gov/web/types";
-import { WarningIcon } from "@app-gov/web/vectors";
 
-import { AccountSelect, Button, IdentityContext } from "./";
+import { Button, IdentityContext } from "./";
 
 interface IdentityFormProps {
 	open: boolean;
@@ -20,8 +14,6 @@ export const IdentityForm: FC<
 	const { twitterRegistrarIndex, discordRegistrarIndex } = useContext(
 		IdentityContext.Context
 	);
-
-	const identityCheck = useIdentityCheck();
 
 	return (
 		<form {...props}>
@@ -44,27 +36,7 @@ export const IdentityForm: FC<
 				stash account that is actively staking, you may connect your controller
 				wallet.
 			</p>
-			<fieldset
-				className={classNames("mb-12 min-w-0", identityCheck && "space-y-4")}
-			>
-				<AccountSelect required name="address" />
-				<If condition={!!identityCheck}>
-					<div className="text-hero float-left inline p-[0.1875rem] pb-0">
-						<WarningIcon className="h-6 w-6" />
-					</div>
-					<p className="prose text-sm leading-7">
-						<If condition={identityCheck?.identitySet}>
-							This account already has a registered identity. Connecting your
-							social channels will overwrite the previously registered channels.
-						</If>
-						<If condition={identityCheck?.judgementProvided}>
-							This account already has judgements provided on its identity.
-							Connecting your social channels will remove the previous
-							judgements.
-						</If>
-					</p>
-				</If>
-			</fieldset>
+			<IdentityContext.Connect />
 
 			<h2 className="font-display border-hero mb-4 border-b-2 text-4xl uppercase">
 				Connect your social channels
@@ -106,38 +78,4 @@ export const IdentityForm: FC<
 			/>
 		</form>
 	);
-};
-
-const useIdentityCheck = () => {
-	const { api } = useCENNZApi();
-	const { selectedAccount } = useCENNZWallet();
-	const [identityCheck, setIdentityCheck] = useState<{
-		identitySet?: boolean;
-		judgementProvided?: boolean;
-	}>();
-
-	useEffect(() => {
-		if (!api || !selectedAccount?.address) return;
-
-		const checkIdentity = async () => {
-			const identity = (await api.query.identity.identityOf(
-				selectedAccount.address
-			)) as Option<IdentityInfo>;
-
-			if (!identity.isSome) return setIdentityCheck(undefined);
-
-			const prevIdentity = identity.toJSON() as unknown as {
-				judgements: AnyTuple[];
-			};
-
-			if (prevIdentity?.judgements.length)
-				return setIdentityCheck({ judgementProvided: true });
-
-			setIdentityCheck({ identitySet: true });
-		};
-
-		checkIdentity();
-	}, [selectedAccount?.address, api]);
-
-	return identityCheck;
 };
